requirements.txt
# Django
django==4.0.4
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
# django-admin-rangefilter==0.10.0
django-cleanup==6.0.0

# Django admin design
django-material-admin==1.8.6

# Django restframework
djangorestframework==3.14.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.1
djangorestframework-api-key==2.2.0
pyjwt==2.6.0
cryptography==39.0.2
pyjwt[crypto]

# Data scientist
celery==5.2.5
redis==4.5.1
numpy==1.23.3
pandas==1.5.0

# Scrapping
requests==2.31.0
beautifulsoup4==4.12.2

# Other
Pillow==9.2.0
psycopg2-binary==2.9.3
gunicorn==20.1.0

date-range-filter==0.0.2
dj-database-url==1.3.0
whitenoise==5.3.0

